backend/requirements.txt
fastapi
uvicorn
sqlmodel
python-multipart
pyjwt
passlib
